Rename: main에 따라 패키지 수정
