+++ src/main/java/com/osundosun/momo/repository/MeetingOwnerRepository.java
@@ -1,7 +1,9 @@
-package com.osundosun.momo.jpa;
+package com.osundosun.momo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.osundosun.momo.entity.MeetingOwner;
 
 @Repository
 public interface MeetingOwnerRepository extends JpaRepository<MeetingOwner, Number>{
